requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.22.2
pygithub==2.4.0
anthropic==0.34.2
bcrypt==4.2.0
reportlab==4.2.2